# docs/swagger.yaml
definitions:
  list.Request:
    properties:
      activeAt:
        type: string
      status:
        type: boolean
      title:
        type: string
    type: object
  response.Object:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /tasks:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Object'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: List of tasks from the database
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: body param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/list.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Add a new task to the database
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: path param
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Delete the task from the database
      tags:
      - tasks
    get:
      consumes:
      - application/json
      parameters:
      - description: path param
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Read the task from the database
      tags:
      - tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: path param
        in: path
        name: id
        required: true
        type: string
      - description: body param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/list.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Update the task in the database
      tags:
      - tasks
  /tasks/{id}/done:
    put:
      consumes:
      - application/json
      parameters:
      - description: path param
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Status the tasks from the database
      tags:
      - tasks
swagger: "2.0"